fixed Cerberus animation speed multiplier

+++ src/main/java/absolutelyaya/ultracraft/accessor/IAnimatedEnemy.java
@@ -6,7 +6,7 @@ public interface IAnimatedEnemy
 	
 	byte getAnimation();
 	
-	float getAnimSpeedMult();
+	int getAnimSpeedMult();
 	
 	void setCooldown(int cooldown);
